requirements.txt
# Core scientific computing
numpy==1.24.4
pandas==2.0.3
scipy==1.10.1

# Optimization
ortools==9.7.2996

# Forecasting
statsmodels==0.14.0
prophet==1.1.5

# Deep learning (for LSTM forecasting)
torch==2.1.0
scikit-learn==1.3.2

# Web frameworks
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.2

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# Data validation
pydantic==2.5.2

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
loguru==0.7.2

# For API
httpx==0.25.2
